feat: Mudanaça de tipos de dados
- Refiz nos seguintes modulos, a alteração de tipo de dados relacionado a precificação, de Float para BigDecimal, e ajustado o mapper da importação.

+++ src/main/java/com/example/ecommerce/modules/cart/dto/CartItemResponseDTO.java
@@ -1,13 +1,14 @@
 package com.example.ecommerce.modules.cart.dto;
 
-import com.example.ecommerce.modules.product.dto.ProductResponseDTO;
 import com.example.ecommerce.modules.product.dto.ProductResponseResumeDTO;
+
+import java.math.BigDecimal;
 
 public record CartItemResponseDTO(
         Long cartItemId,
         Integer quantidade,
-        Float precoUnitario,
-        Float subtotal,
+        BigDecimal precoUnitario,
+        BigDecimal subtotal,
         ProductResponseResumeDTO products
 ) {
 }
