+++ src/main/java/com/example/ecommerce/modules/product/dto/ProductCreateDTO.java
@@ -2,13 +2,15 @@ package com.example.ecommerce.modules.product.dto;
 
 import com.example.ecommerce.modules.product.model.ProductEnum;
 
+import java.math.BigDecimal;
+
 public record ProductCreateDTO(
         String nome,
         String slug,
         String descricaoCurta,
         String descricao,
-        Float preco,
-        Float precoPromocional,
+        BigDecimal preco,
+        BigDecimal precoPromocional,
         Integer quantidadeEstoque,
         Integer estoqueMinimo,
         String sku,
