+++ src/main/java/com/example/ecommerce/modules/product/dto/ProductResponseDTO.java
@@ -2,6 +2,7 @@ package com.example.ecommerce.modules.product.dto;
 
 import com.example.ecommerce.modules.product.model.ProductEnum;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record ProductResponseDTO(
@@ -10,17 +11,17 @@ public record ProductResponseDTO(
         String slug,
         String descricaoCurta,
         String descricao,
-        float preco,
-        float precoPromocional,
+        BigDecimal preco,
+        BigDecimal precoPromocional,
         Integer quantidadeEstoque,
         Integer quantidadeReservada,
         Integer estoqueMinimo,
         String sku,
-        float peso,
-        float altura,
-        float largura,
-        float comprimento,
-        float mediaAvaliacao,
+        Float peso,
+        Float altura,
+        Float largura,
+        Float comprimento,
+        Float mediaAvaliacao,
         Integer totalAvaliacoes,
         ProductEnum status,
         LocalDate dataCriacao,
